feat: land Sol-RL post-training workflow updates
Move the post-training stack under diffusion/post_training and wire in the launch, checkpoint, dataset, and documentation updates needed to run Sol-RL directly from the Sana repo.
This also surfaces Sol-RL in the main README and docs so the NVlabs branch is ready to share and use.

Made-with: Cursor
