+++ configs/sol_rl/Sana1.0_1600M_linear.yaml
@@ -17,11 +17,11 @@ data:
   sort_dataset: false
 # model config
 model:
-  model: SanaMSLinear_1600M_P1_D20
+  model: SanaMSLinearFFN_1600M_P1_D20
   image_size: 1024
   mixed_precision: bf16
   fp32_attention: true
-  load_from: sana10_native_linear.pth
+  load_from: output/pretrained_models/SANA_LinearFFN.pth
   resume_from:
   aspect_ratio_type: ASPECT_RATIO_1024
   multi_scale: true
