--- diffusion/flow_grpo/pickscore_scorer.py
@@ -1,85 +1,0 @@
-import torch
-from PIL import Image
-from transformers import AutoModel, AutoProcessor
-
-
-class PickScoreScorer(torch.nn.Module):
-    def __init__(self, device="cuda", dtype=torch.float32):
-        super().__init__()
-        processor_path = "laion/CLIP-ViT-H-14-laion2B-s32B-b79K"
-        model_path = "yuvalkirstain/PickScore_v1"
-        self.device = device
-        self.dtype = dtype
-        self.processor = AutoProcessor.from_pretrained(processor_path)
-        self.model = AutoModel.from_pretrained(model_path).eval().to(device)
-        self.model = self.model.to(dtype=dtype)
-
-    @staticmethod
-    def _as_embedding(features):
-        """Convert model output to [B, C] embedding tensor."""
-        if torch.is_tensor(features):
-            return features
-
-        # transformers BaseModelOutputWithPooling
-        if hasattr(features, "pooler_output") and features.pooler_output is not None:
-            return features.pooler_output
-        if hasattr(features, "last_hidden_state") and features.last_hidden_state is not None:
-            hidden = features.last_hidden_state
-            if hidden.ndim == 3:
-                return hidden.mean(dim=1)
-            return hidden
-
-        raise TypeError(f"Unsupported model output type: {type(features)}")
-
-    @torch.no_grad()
-    def __call__(self, prompt, images):
-        # Preprocess images
-        image_inputs = self.processor(
-            images=images,
-            padding=True,
-            truncation=True,
-            max_length=77,
-            return_tensors="pt",
-        )
-        image_inputs = {k: v.to(device=self.device) for k, v in image_inputs.items()}
-        # Preprocess text
-        text_inputs = self.processor(
-            text=prompt,
-            padding=True,
-            truncation=True,
-            max_length=77,
-            return_tensors="pt",
-        )
-        text_inputs = {k: v.to(device=self.device) for k, v in text_inputs.items()}
-
-        # Get embeddings
-        image_embs = self._as_embedding(self.model.get_image_features(**image_inputs))
-        image_embs = image_embs / image_embs.norm(p=2, dim=-1, keepdim=True)
-
-        text_embs = self._as_embedding(self.model.get_text_features(**text_inputs))
-        text_embs = text_embs / text_embs.norm(p=2, dim=-1, keepdim=True)
-
-        # Calculate scores
-        logit_scale = self.model.logit_scale.exp()
-        scores = logit_scale * (text_embs @ image_embs.T)
-        scores = scores.diag()
-        # norm到0-1
-        scores = scores / 26
-        return scores
-
-
-# Usage example
-def main():
-    scorer = PickScoreScorer(device="cuda", dtype=torch.float32)
-    images = [
-        "test_cases/nasa.jpg",
-    ]
-    pil_images = [Image.open(img) for img in images]
-    prompts = [
-        'An astronaut’s glove floating in zero-g with "NASA 2049" on the wrist',
-    ]
-    print(scorer(prompts, pil_images))
-
-
-if __name__ == "__main__":
-    main()
