--- diffusion/flow_grpo/reward_ckpt_path.py
@@ -1,3 +1,0 @@
-import os
-
-CKPT_PATH = os.path.join(os.path.dirname(os.path.abspath(__file__)), "../../reward_ckpts")
