+++ diffusion/model/builder.py
@@ -20,9 +20,8 @@ import tempfile
 import numpy as np
 import torch
 import torch.nn.functional as F
-from diffusers import AutoencoderDC
+from diffusers import AutoencoderDC, AutoencoderKLLTXVideo
 from diffusers.models import AutoencoderKL
-from diffusers.models.autoencoders import AutoencoderKLLTX2Video
 from mmcv import Registry
 from termcolor import colored
 from transformers import (
@@ -177,11 +176,11 @@ def get_vae(name, model_path, device="cuda", dtype=None, config=None):
         )
         return vae
     elif "LTX2VAE_diffusers" in name:
-        # Use diffusers AutoencoderKLLTX2Video for LTX2
+        # Use diffusers AutoencoderKLLTXVideo for LTX2
         assert config is not None, "config.vae is required for LTX2VAE_diffusers"
         print(colored(f"[LTX2VAE_diffusers] Loading model from {config.vae_pretrained}", attrs=["bold"]))
         vae = (
-            AutoencoderKLLTX2Video.from_pretrained(config.vae_pretrained, subfolder="vae", torch_dtype=dtype)
+            AutoencoderKLLTXVideo.from_pretrained(config.vae_pretrained, subfolder="vae", torch_dtype=dtype)
             .to(device)
             .eval()
         )
@@ -304,7 +303,7 @@ def vae_encode(name, vae, images, sample_posterior=True, device="cuda", cache_ke
         z = ae.encode(images.to(device))
         z = torch.stack(z, dim=0)
     elif "LTX2VAE_diffusers" in name:
-        # LTX2VAE_diffusers (diffusers AutoencoderKLLTX2Video) expects input shape (B, C, T, H, W) with value range [-1, 1]
+        # LTX2VAE_diffusers (diffusers AutoencoderKLLTXVideo) expects input shape (B, C, T, H, W) with value range [-1, 1]
         posterior = vae.encode(images.to(device=vae.device, dtype=vae.dtype)).latent_dist
         z = posterior.mode()
         # Normalize latents: z = (z - mean) / std * scaling_factor
@@ -356,7 +355,7 @@ def vae_decode(name, vae, latent):
     elif "Wan2_2_VAE" in name:
         samples = vae.decode(latent)
     elif "LTX2VAE_diffusers" in name:
-        # LTX2VAE_diffusers (diffusers AutoencoderKLLTX2Video)
+        # LTX2VAE_diffusers (diffusers AutoencoderKLLTXVideo)
         # Denormalize latents: z = z * std / scaling_factor + mean
         latents_mean = vae.latents_mean.view(1, -1, 1, 1, 1).to(latent.device, latent.dtype)
         latents_std = vae.latents_std.view(1, -1, 1, 1, 1).to(latent.device, latent.dtype)
