+++ diffusion/post_training/dataset/drawbench/train.txt
@@ -1,0 +1,1 @@
+An oil painting of a couple in formal evening wear going home get caught in a heavy downpour with no umbrellas.
