+++ mkdocs.yml
@@ -78,10 +78,12 @@ nav:
       - SANA-Video: sana_video.md
       - Video Inference: sana_video_inference.md
       - LongSANA: longsana.md
+  - Post Training:
+      - Cosmos-RL: sana_cosmos_rl.md
+      - Sol-RL: sol_rl.md
   - Advanced:
       - ControlNet: sana_controlnet.md
       - LoRA & DreamBooth: sana_lora_dreambooth.md
-      - Cosmos-RL: sana_cosmos_rl.md
   - Integrations:
       - ComfyUI: ComfyUI/comfyui.md
       - SGLang: sglang.md
